Add files via upload
Foresight Pharmaceuticals is one of the leading Pharmaceutical Manufacturing companies with a global presence.
Their Markets are divided into different regions across the world. One of those regions manages the Germany and Poland Markets.
But the company does not sell directly to customers. Instead, they work with a couple of Distributors in all their regions.
They have an agreement with each of the distributor to share their Sales Data with them. This is to enable them gain insights up to the retail level.